+++ source/modules/MailMerge/views/createMergeLead.php
@@ -17,6 +17,7 @@ class MailMerge_createMergeLead_View extends Vtiger_Index_View {
 
     public function process(Vtiger_Request $request) {
 // Fix For: http://trac.vtiger.com/cgi-bin/trac.cgi/ticket/2107
+        $default_charset= $GLOBALS['default_charset'];
         $randomfilename = "vt_" . str_replace(array(".", " "), "", microtime());
 
         $templateid = $_REQUEST['document'];
@@ -243,7 +244,7 @@ else if(document.all)
     //delete old .odt files in the wordtemplatedownload directory
    foreach (glob("$path*") as $delefile) 
     {
-        if(GetFileExtension($delefile)!='zip'){
+        if(GetFileExtension($delefile)!='zip' && !is_dir($delefile) && file_exists($delefile)){
         unlink($delefile);
         }
       
@@ -359,7 +360,9 @@ else if(document.all)
                 remove_dir($wordtemplatedownloadpath . $temp_dir);
             }
             $zip = new ZipArchive;
+            if(file_exists($wordtemplatedownloadpath . $filename)){
             unlink($wordtemplatedownloadpath . $filename);
+            }
             $zip->open($path . $randomfilename . '.zip', ZipArchive::CREATE);
             foreach (glob($wordtemplatedownloadpath . "*.odt") as $file) {
                 $zip->addFile($file, $file);
@@ -415,7 +418,7 @@ else if(document.all)
     //delete old .odt files in the wordtemplatedownload directory
    foreach (glob("$path*") as $delefile) 
     {
-        if(GetFileExtension($delefile)!='zip'){
+        if(GetFileExtension($delefile)!='zip' && !is_dir($delefile) && file_exists($delefile)){
         unlink($delefile);
         }
        
@@ -516,7 +519,9 @@ else if(document.all)
 
             }
              $zip = new ZipArchive;
+             if(file_exists($wordtemplatedownloadpath . $filename)){
             unlink($wordtemplatedownloadpath . $filename);
+             }
             $zip->open($path . $randomfilename . '.zip', ZipArchive::CREATE);
             foreach (glob($wordtemplatedownloadpath . "*.rtf") as $file) {
                 $zip->addFile($file, $file);
